montly sales report kinda works
