+++ app/routes.py
@@ -455,142 +455,6 @@ def payment_report():
     except Exception as e:
         return jsonify({'error': str(e)}), 500
 
-'''This Route is just breaking right now before even testing with /forward'''
-@app.route('/api/manager/monthly-sales-report', methods=['GET'])
-# this API generates monthly sales reports on all payments made in the dealership
-# Scufffedd but i think it works
-def monthly_sales_report():
-    # checks if a manager is logged in to view the information | uncomment when we have it working for sure 100% on frontend in implementation | commented manager session out for testing and ensuring it works
-    # employee_id = session.get('employee_session_id')
-    # if not employee_id:
-    #     return jsonify({'message': 'Unauthorized access'}), 401
-    #
-    # ensures that the employee is a Technician
-    # employee = Employee.query.filter_by(employeeID=employee_id, employeeType='Manager').first()
-    # if not employee:
-    #     return jsonify({'message': 'Unauthorized access'}), 401
-
-    # request the parameters for monthly reports and in which year as well
-    # have these values return in some like tab like or drop down value from the manager dashboard view
-    # Get month and year from request parameters
-    month = request.args.get('month')  # 1-12 for selected month
-    year = request.args.get('year')    # YYYY format for selected year
-
-    # Validate month and year inputs
-    if not month or not year:
-        return jsonify({'message': 'Month and year are required parameters'}), 400
-
-    try:
-        month = int(month)
-        year = int(year)
-    except ValueError:
-        return jsonify({'message': 'Invalid month or year format'}), 400
-
-    # Validate month range
-    if month < 1 or month > 12:
-        return jsonify({'message': 'Invalid month value. Month should be between 1 and 12'}), 400
-
-
-    # Debugging statements for month and year
-    print("Selected month:", month)
-    print("Selected year:", year)
-
-    # Get current date
-    current_date = datetime.now()
-
-    # Debugging statement for current date
-    print("Current date:", current_date)
-
-    # Get the last day of the selected month dynamically
-    last_day_of_month = calendar.monthrange(int(year), int(month))[1]
-
-    # Set the start date and end date using the last day of the month
-    start_date = datetime(int(year), int(month), 1)
-    end_date = datetime(int(year), int(month), last_day_of_month)
-
-    # Hardcoded definitions for dates
-    start_of_year = datetime(int(year), 1, 1)
-    last_year_start = datetime(int(year) - 1, 1, 1)
-    last_year_end = datetime(int(year) - 1, 12, 31)
-    last_year_month_start = datetime(int(year) - 1, int(month), 1)
-    
-    # Check if it's February and the last day of the month is 29 (indicating a leap year)
-    if int(month) == 2 and last_day_of_month == 29:
-        # Adjust the last day of February to 28 for non-leap years
-        last_day_of_month = 28
-        
-    last_year_month_end = datetime(int(year) - 1, int(month), last_day_of_month)
-
-    # Debugging statements for dates
-    print("Start date of selected month:", start_date)
-    print("End date of selected month:", end_date)
-    print("Start date of last year:", last_year_start)
-    print("End date of last year:", last_year_end)
-    print("Start date of last year's selected month:", last_year_month_start)
-    print("End date of last year's selected month:", last_year_month_end)
-    # Query purchases for monthly report
-    purchases = Purchases.query.filter(db.extract('month', Purchases.purchaseDate) == month,
-                                       db.extract('year', Purchases.purchaseDate) == year).all()
-
-    # Prepare sales report data for monthly report
-    total_sales = 0
-    sales_report = []
-
-    # Calculate total sales and populate sales report for monthly report
-    for purchase in purchases:
-        bid_value = 0  # Default value if bid is None
-        bid_id = None  # Default value for bid ID
-        if purchase.bidID is not None:
-            # If a bid is associated with the purchase, retrieve the bid value and ID
-            bid = Bids.query.get(purchase.bidID)
-            if bid:
-                bid_value = bid.bidValue
-                bid_id = bid.bidID
-            else:
-                print(f"Bid not found for Purchase ID: {purchase.purchaseID}")
-
-        total_sales += bid_value
-
-        sales_report.append({
-            'purchase_id': purchase.purchaseID,
-            'member_id': purchase.memberID,
-            'vehicle_id': purchase.VIN_carID,
-            'confirmation_number': purchase.confirmationNumber,
-            'purchase_type': purchase.purchaseType,
-            'purchase_timestamp': purchase.purchaseDate.isoformat(),
-            'bid_id': bid_id,
-            'bid_value': str(bid_value)  # Convert Decimal to string for JSON serialization
-        })
-
-    # Query purchases for other types of reports
-    all_time_sales = Purchases.query.all()
-    yearly_sales = Purchases.query.filter(Purchases.purchaseDate >= start_date, Purchases.purchaseDate <= end_date).all()
-    last_year_sales = Purchases.query.filter(Purchases.purchaseDate >= last_year_start, Purchases.purchaseDate <= last_year_end).all()
-    last_year_month_sales = Purchases.query.filter(Purchases.purchaseDate >= last_year_month_start, Purchases.purchaseDate <= last_year_month_end).all()
-
-    yearly_purchases = Purchases.query.filter(db.extract('year', Purchases.purchaseDate) == year).all()
-
-    # Calculate total sales for other types of reports
-    total_all_time_sales = sum(Bids.query.get(purchase.bidID).bidValue if (purchase.bidID is not None and Bids.query.get(purchase.bidID) is not None) else 0 for purchase in all_time_sales)
-    total_yearly_sales = sum(Bids.query.get(purchase.bidID).bidValue if (purchase.bidID is not None and Bids.query.get(purchase.bidID) is not None) else 0 for purchase in yearly_sales)
-    total_last_year_sales = sum(Bids.query.get(purchase.bidID).bidValue if (purchase.bidID is not None and Bids.query.get(purchase.bidID) is not None) else 0 for purchase in last_year_sales)
-    total_last_year_month_sales = sum(Bids.query.get(purchase.bidID).bidValue if (purchase.bidID is not None and Bids.query.get(purchase.bidID) is not None) else 0 for purchase in last_year_month_sales)
-
-    # Calculate total sales for the current year
-    total_this_year = sum(Bids.query.get(purchase.bidID).bidValue if (purchase.bidID is not None and Bids.query.get(purchase.bidID) is not None) else 0 for purchase in yearly_purchases)
-
-
-
-    return jsonify({
-        'total_sales': str(total_sales),  # Convert total sales to string for JSON serialization
-        'total_this_year': str(total_this_year),  # Total sales for the selected year
-        'all_time_sales': str(total_all_time_sales),
-        'yearly_sales': str(total_yearly_sales),
-        'last_year_sales': str(total_last_year_sales),
-        'last_year_month_sales': str(total_last_year_month_sales),
-        'sales_report': sales_report
-    }), 200
-
 
 
 @app.route('/api/customer/make-payment', methods=['POST'])
